Moved DrawPoint composition to series folder.
