+++ ts/Series/DrawPointComposition.ts
